+++ zfit/minimizers/errors.py
@@ -1,6 +1,7 @@
 #  Copyright (c) 2020 zfit
 
 import numpy as np
+import tensorflow as tf
 from scipy import optimize
 
 from ..param import set_values
@@ -12,147 +13,11 @@ class NewMinimum(Exception):
     pass
 
 
-def pll(minimizer, loss, params, values) -> 'FitResult':
-    """Compute minimum profile likelihood for given parameters and values."""
-    params = convert_to_container(params)
-    values = convert_to_container(values)
-
-    verbosity = minimizer.verbosity
-    minimizer.verbosity = 0
-
-    with set_values(params, values):
-        for param in params:
-            param.floating = False
-
-        minimum = minimizer.minimize(loss=loss)
-
-    for param in params:
-        param.floating = True
-    minimizer.verbosity = verbosity
-
-    return minimum
+class FailEvalLossNaN(Exception):
+    pass
 
 
-def get_crossing_value(result, params, direction, sigma, rootf, rtol):
-    """Find the crossing point between the profiled loss function, for given parameters, and the value of
-    `errordef` for a given direction (positive / negative).
-    `errordef` = 1 for a chisquare fit, = 0.5 for a likelihood fit.
-    """
-
-    all_params = list(result.params.keys())
-    loss = result.loss
-    errordef = loss.errordef
-    fmin = result.fmin
-    minimizer = result.minimizer.copy()
-    minimizer.tolerance = minimizer.tolerance * 0.5
-    rtol *= errordef
-
-    set_values(all_params, result)
-
-    covariance = result.covariance(as_dict=True)
-
-    step_size_dict = {ap: ap.step_size for ap in all_params}
-    for ap in all_params:
-        ap.step_size = covariance[(ap, ap)] ** 0.5
-
-    to_return = {}
-    for param in params:
-        param_error = result.hesse(params=param)[param]["error"]
-        param_value = result.params[param]["value"]
-        exp_root = param_value + sigma * direction * param_error  # expected root
-
-        for ap in all_params:
-            if ap == param:
-                continue
-
-            # shift parameters, other than param, using covariance matrix
-            ap_value = result.params[ap]["value"]
-            ap_value += sigma * direction * covariance[(param, ap)] * (2 * errordef / param_error ** 2) ** 0.5
-            ap.set_value(ap_value)
-
-        cache = {}  # TODO: round for better floating point comparison?
-
-        def shifted_pll(v):
-            """
-            Computes the pll, with the minimum substracted and shifted by minus the `errordef`, for a
-            given parameter.
-            `errordef` = 1 for a chisquare fit, = 0.5 for a likelihood fit.
-
-            The function raises a `NewMinimum` exception if the value of the shifted `pll` is less than
-            `- errordef`.
-            """
-            if v not in cache:
-                pll_result = pll(minimizer, loss, param, v)
-                cache[v] = pll_result.fmin - fmin - errordef
-                if cache[v] < -errordef - minimizer.tolerance:
-                    set_values(pll_result.params.keys(), pll_result)  # set values to the new minimum
-                    raise NewMinimum("A new is minimum found.")
-
-            return cache[v]
-
-        exp_shifted_pll = shifted_pll(exp_root)
-
-        if np.allclose(0., exp_shifted_pll, atol=0.0005):
-            root = exp_root
-
-        else:
-            def linear_interp(y):
-                """
-                Linear interpolation between the minimum of the `shifted_pll` curve and its expected root,
-                assuming it is a parabolic curve.
-                """
-                slope = (exp_root - param_value) / (exp_shifted_pll + errordef)
-                return param_value + (y + errordef) * slope
-
-            bound_interp = linear_interp(0)
-
-            if exp_shifted_pll > 0.:
-                lower_bound = exp_root
-                upper_bound = bound_interp
-            else:
-                lower_bound = bound_interp
-                upper_bound = exp_root
-
-            if direction == 1:
-                lower_bound, upper_bound = upper_bound, lower_bound
-
-            # Check if the `shifted_pll` function has the same sign at the lower and upper bounds.
-            # If they have the same sign, the window given to the root finding algorithm is increased.
-
-            nsigma = 1.5
-            while np.sign(shifted_pll(lower_bound)) == np.sign(shifted_pll(upper_bound)):
-                if direction == -1:
-                    if np.sign(shifted_pll(lower_bound)) == -1:
-                        lower_bound = param_value - nsigma * param_error
-                    else:
-                        upper_bound = param_value
-                else:
-                    if np.sign(shifted_pll(lower_bound)) == -1:
-                        upper_bound = param_value + nsigma * param_error
-                    else:
-                        lower_bound = param_value
-
-                nsigma += 0.5
-
-            root, results = rootf(f=shifted_pll, a=lower_bound, b=upper_bound, rtol=rtol,
-                                  full_output=True)
-
-        to_return[param] = root
-
-    for ap in all_params:
-        ap.step_size = step_size_dict[ap]
-
-    return to_return
-
-
-def _rootf(**kwargs):
-    return optimize.toms748(k=1, **kwargs)
-
-
-# def _rootf(**kwargs):
-#     return optimize.brentq(**kwargs)
-
-def compute_errors(result, params, sigma=1, rootf=_rootf, rtol=0.01):
+def compute_errors(result, params, sigma=1, rtol=0.001, method="hybr", covariance_method=None):
     """
     Computes asymmetric errors of parameters by profiling the loss function in the fit result.
 
@@ -161,8 +26,10 @@ def compute_errors(result, params, sigma=1, rootf=_rootf, rtol=0.01):
         params (list(:py:class:`~zfit.Parameter`)): The parameters to calculate the
             errors error. If None, use all parameters.
         sigma (float): Errors are calculated with respect to `sigma` std deviations.
-        rootf (callable): function used to find the roots of the loss function
         rtol (float, default=0.01): relative tolerance between the computed and the exact roots
+        method (str, defautl='hybr'): type of solver, `method` argument of `scipy.optimize.root_
+        covariance_method (str or Callable): The method to use to calculate the correlation matrix. Valid choices are
+            {'minuit_hesse', 'hesse_np'} or a Callable.
 
     Returns:
         `OrderedDict`: A `OrderedDict` containing as keys the parameter and as value a `dict` which
@@ -174,19 +41,57 @@ def compute_errors(result, params, sigma=1, rootf=_rootf, rtol=0.01):
     params = convert_to_container(params)
     new_result = None
 
+    all_params = list(result.params.keys())
+    loss = result.loss
+    errordef = loss.errordef
+    fmin = result.fmin
+    rtol *= errordef
+    minimizer = result.minimizer
+
+    covariance = result.covariance(method=covariance_method, as_dict=True)
+    set_values(all_params, result)
+
     try:
-
-        upper_values = get_crossing_value(result=result, params=params, direction=1, sigma=sigma,
-                                          rootf=rootf, rtol=rtol)
-
-        lower_values = get_crossing_value(result=result, params=params, direction=-1, sigma=sigma,
-                                          rootf=rootf, rtol=rtol)
 
         to_return = {}
         for param in params:
-            fitted_value = result.params[param]["value"]
-            to_return[param] = {"lower": lower_values[param] - fitted_value,
-                                "upper": upper_values[param] - fitted_value}
+            param_error = covariance[(param, param)]**0.5
+            param_value = result.params[param]["value"]
+            other_params = [p for p in all_params if p != param]
+
+            initial_values = {"lower": [], "upper": []}
+            direction = {"lower": -sigma, "upper": sigma}
+
+            for ap in all_params:
+                for d in ["lower", "upper"]:
+                    ap_value = result.params[ap]["value"]
+                    ap_value += direction[d] * covariance[(param, ap)] * (2 * errordef / param_error ** 2) ** 0.5
+                    initial_values[d].append(ap_value)
+
+            def func(values):
+
+                with set_values(all_params, values):
+                    try:
+                        loss_value, gradients = loss.value_gradients(params=other_params)
+                    except tf.errors.InvalidArgumentError:
+                        msg = (f"The evaluation of the errors of {param.name} failed due to too many NaNs"
+                               " being produced in the loss and/or its gradients. This is most probably"
+                               " caused by negative values returned from the PDF.")
+                        raise FailEvalLossNaN(msg)
+
+                    shifted_loss = loss_value.numpy() - fmin - errordef * sigma**2
+                    gradients = np.array(gradients)
+
+                if shifted_loss < -errordef - minimizer.tolerance:
+                    set_values(all_params, values)  # set values to the new minimum
+                    raise NewMinimum("A new is minimum found.")
+
+                return np.concatenate([[shifted_loss], gradients])
+
+            to_return[param] = {}
+            for d in ["lower", "upper"]:
+                roots = optimize.root(fun=func, x0=initial_values[d], tol=rtol, method=method)
+                to_return[param][d] = roots.x[all_params.index(param)] - param_value
 
     except NewMinimum as e:
         from .. import settings
@@ -199,8 +104,8 @@ def compute_errors(result, params, sigma=1, rootf=_rootf, rtol=0.01):
         if new_result.fmin >= new_found_fmin:
             raise RuntimeError("A new minimum was discovered but the minimizer was not able to find this on himself. "
                                "This behavior is currently an exception but will most likely change in the future.")
-        to_return, new_result_ = compute_errors(result=new_result, params=params, sigma=sigma,
-                                                rootf=rootf, rtol=rtol)
+        to_return, new_result_ = compute_errors(result=new_result, params=params, sigma=sigma, rtol=rtol,
+                                                method=method)
         if new_result_ is not None:
             new_result = new_result_